+++ tools/test-js-with-node
@@ -54,6 +54,7 @@ EXEMPT_FILES = make_set(
         "web/src/archive.js",
         "web/src/assets.d.ts",
         "web/src/attachments_ui.ts",
+        "web/src/audible_notifications.js",
         "web/src/avatar.ts",
         "web/src/billing/event_status.ts",
         "web/src/billing/helpers.ts",
