+++ web/src/notifications.js
@@ -1,7 +1,5 @@
 import $ from "jquery";
 
-import {user_settings} from "./user_settings";
-
 export const notice_memory = new Map();
 
 export let NotificationAPI;
@@ -50,25 +48,6 @@ export function initialize() {
         }
         notice_memory.clear();
     });
-
-    update_notification_sound_source($("#user-notification-sound-audio"), user_settings);
-}
-
-export function update_notification_sound_source(container_elem, settings_object) {
-    const notification_sound = settings_object.notification_sound;
-    const audio_file_without_extension = "/static/audio/notification_sounds/" + notification_sound;
-    container_elem
-        .find(".notification-sound-source-ogg")
-        .attr("src", `${audio_file_without_extension}.ogg`);
-    container_elem
-        .find(".notification-sound-source-mp3")
-        .attr("src", `${audio_file_without_extension}.mp3`);
-
-    if (notification_sound !== "none") {
-        // Load it so that it is ready to be played; without this the old sound
-        // is played.
-        container_elem[0].load();
-    }
 }
 
 export function permission_state() {
