+++ web/src/server_events_dispatch.js
@@ -4,6 +4,7 @@ import * as activity_ui from "./activity_ui";
 import * as alert_words from "./alert_words";
 import * as alert_words_ui from "./alert_words_ui";
 import * as attachments_ui from "./attachments_ui";
+import * as audible_notifications from "./audible_notifications";
 import * as blueslip from "./blueslip";
 import * as bot_data from "./bot_data";
 import * as browser_history from "./browser_history";
@@ -30,7 +31,6 @@ import * as muted_users_ui from "./muted_users_ui";
 import * as narrow_state from "./narrow_state";
 import * as narrow_title from "./narrow_title";
 import * as navbar_alerts from "./navbar_alerts";
-import * as notifications from "./notifications";
 import * as overlays from "./overlays";
 import {page_params} from "./page_params";
 import * as peer_data from "./peer_data";
@@ -450,7 +450,7 @@ export function dispatch_normal_event(event) {
             settings_realm_user_settings_defaults.update_page(event.property);
 
             if (event.property === "notification_sound") {
-                notifications.update_notification_sound_source(
+                audible_notifications.update_notification_sound_source(
                     $("#realm-default-notification-sound-audio"),
                     realm_user_settings_defaults,
                 );
@@ -673,7 +673,7 @@ export function dispatch_normal_event(event) {
 
                 if (notification_name === "notification_sound") {
                     // Change the sound source with the new page `notification_sound`.
-                    notifications.update_notification_sound_source(
+                    audible_notifications.update_notification_sound_source(
                         $("#user-notification-sound-audio"),
                         user_settings,
                     );
